refactor: rebuild structure completely (#24)
