+++ package.json
@@ -2,10 +2,10 @@
 	"name": "quiztime-react-backend",
 	"version": "1.0.0",
 	"description": "",
-	"main": "index.js",
+	"main": "src/index.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node index.js"
+		"start": "node src/index.js"
 	},
 	"keywords": [],
 	"author": "",
